improved pagination & structure
