--- src/components/Pagination.tsx
@@ -1,48 +1,0 @@
-"use client";
-
-import React from "react";
-import { Button } from "@/components/ui/button";
-
-type PaginationProps = {
-  page: number;
-  pageSize: number;
-  total: number;
-  onPageChange: (page: number) => void;
-};
-
-export default function Pagination({
-  page,
-  pageSize,
-  total,
-  onPageChange,
-}: PaginationProps) {
-  const totalPages = Math.max(1, Math.ceil(total / pageSize));
-  const canPrev = page > 1;
-  const canNext = page < totalPages;
-
-  return (
-    <div className="flex flex-col gap-3 sm:flex-row sm:items-center sm:justify-between">
-      <div className="text-sm text-muted-foreground text-center sm:text-left">
-        Page {page} of {totalPages}
-      </div>
-      <div className="flex gap-2 justify-center sm:justify-end">
-        <Button
-          variant="outline"
-          size="sm"
-          disabled={!canPrev}
-          onClick={() => onPageChange(page - 1)}
-        >
-          Prev
-        </Button>
-        <Button
-          variant="outline"
-          size="sm"
-          disabled={!canNext}
-          onClick={() => onPageChange(page + 1)}
-        >
-          Next
-        </Button>
-      </div>
-    </div>
-  );
-}
