+++ src/components/headers/CurrencySwitcher.tsx
