+++ src/components/sparkline/index.ts
@@ -1,0 +1,2 @@
+export { Sparkline } from "./Sparkline";
+export { SymbolSparkline } from "./SymbolSparkline";
