+++ src/components/table/Pagination.tsx
@@ -1,0 +1,111 @@
+"use client";
+
+import React from "react";
+import { Button } from "../ui/button";
+
+type PaginationProps = {
+  page: number;
+  pageSize: number;
+  total: number;
+  onPageChange: (page: number) => void;
+};
+
+export const Pagination = ({
+  page,
+  pageSize,
+  total,
+  onPageChange,
+}: PaginationProps) => {
+  const totalPages = Math.max(1, Math.ceil(total / pageSize));
+  const canPrev = page > 1;
+  const canNext = page < totalPages;
+
+  const getPageNumbers = () => {
+    const pages: (number | string)[] = [];
+    const maxVisiblePages = 5;
+
+    if (totalPages <= maxVisiblePages) {
+      for (let i = 1; i <= totalPages; i++) {
+        pages.push(i);
+      }
+    } else {
+      pages.push(1);
+
+      if (page <= 3) {
+        for (let i = 2; i <= Math.min(5, totalPages - 1); i++) {
+          pages.push(i);
+        }
+        if (totalPages > 5) {
+          pages.push("...");
+        }
+      } else if (page >= totalPages - 2) {
+        if (totalPages > 5) {
+          pages.push("...");
+        }
+        for (let i = Math.max(totalPages - 4, 2); i <= totalPages - 1; i++) {
+          pages.push(i);
+        }
+      } else {
+        pages.push("...");
+        for (let i = page - 1; i <= page + 1; i++) {
+          pages.push(i);
+        }
+        pages.push("...");
+      }
+
+      if (totalPages > 1) {
+        pages.push(totalPages);
+      }
+    }
+
+    return pages;
+  };
+
+  const pageNumbers = getPageNumbers();
+
+  return (
+    <div className="flex flex-col gap-3 sm:flex-row sm:items-center sm:justify-between">
+      <div className="text-sm text-muted-foreground text-center sm:text-left">
+        Page {page} of {totalPages}
+      </div>
+      <div className="flex gap-1 justify-center sm:justify-end">
+        <Button
+          variant="outline"
+          size="sm"
+          disabled={!canPrev}
+          onClick={() => onPageChange(page - 1)}
+        >
+          Prev
+        </Button>
+
+        {pageNumbers.map((pageNum, index) => (
+          <React.Fragment key={index}>
+            {typeof pageNum === "number" ? (
+              <Button
+                variant={pageNum === page ? "default" : "outline"}
+                size="sm"
+                onClick={() => onPageChange(pageNum)}
+                className="min-w-[2.5rem]"
+              >
+                {pageNum}
+              </Button>
+            ) : (
+              <span className="px-2 py-1 text-sm text-muted-foreground">
+                ...
+              </span>
+            )}
+          </React.Fragment>
+        ))}
+
+        <Button
+          variant="outline"
+          size="sm"
+          disabled={!canNext}
+          onClick={() => onPageChange(page + 1)}
+        >
+          Next
+        </Button>
+      </div>
+    </div>
+  );
+};
