+++ src/components/table/index.ts
@@ -1,0 +1,9 @@
+export { MarketTable } from "./marketTable/MarketTable";
+export { MarketTableBody } from "./marketTable/MarketTableBody";
+export { MarketTableFilters } from "./marketTable/MarketTableFilters";
+export { MarketTableHeader } from "./marketTable/MarketTableHeader";
+export { MarketTableRow } from "./marketTable/MarketTableRow";
+export { MarketTableSkeleton } from "./marketTable/MarketTableSkeleton";
+export { Pagination } from "./Pagination";
+export { SearchBox } from "./SearchBox";
+export { CoinDetailModal } from "./CoinDetailModal";
