+++ src/types/external.d.ts
@@ -1,3 +1,6 @@
 declare module "framer-motion" {
-  export const motion: any;
+  export const motion: {
+    div: React.ComponentType<React.HTMLAttributes<HTMLDivElement>>;
+    [key: string]: React.ComponentType<React.HTMLAttributes<HTMLElement>>;
+  };
 }
